Quick and dirty partial ImGui integration

+++ D3D12/DynamicResourceAllocator.cpp
@@ -58,7 +58,7 @@ DynamicAllocation DynamicResourceAllocator::Allocate(int size)
 	return allocation;
 }
 
-void DynamicResourceAllocator::Free(int fenceValue)
+void DynamicResourceAllocator::Free(uint64 fenceValue)
 {
 	while (m_FenceOffsets.size() > 0)
 	{
